Start history watch on shortcuts access
