+++ apps/extension/src/components/StoreListener.tsx
@@ -1,27 +1,10 @@
 import { setPersonsInStorage } from '@/PersonsPanel/utils';
+import { startHistoryWatch } from '@/utils/history';
 import { IUpdateTaggedPerson } from '@bypass/shared';
 import { getPersons } from '@helpers/fetchFromStorage';
 import useHistoryStore from '@store/history';
 import usePersonStore from '@store/person';
 import { memo, useEffect } from 'react';
-
-const THIRTY_SECONDS = 30 * 1000; // in milliseconds
-
-const isHistoryAlreadyActive = async () => {
-  const { historyStartTime } = await chrome.storage.local.get([
-    'historyStartTime',
-  ]);
-  return Boolean(historyStartTime);
-};
-
-export const startHistoryWatch = async () => {
-  if (await isHistoryAlreadyActive()) {
-    return;
-  }
-  await chrome.storage.local.set({
-    historyStartTime: Date.now() - THIRTY_SECONDS,
-  });
-};
 
 const getSeparatedPersons = (prevPersons: string[], newPersons: string[]) => {
   const prevPersonsSet = new Set(prevPersons);
